projects(tests): refactor handlers tests
