+++ rdmo/projects/tests/conftest.py
@@ -1,0 +1,15 @@
+import pytest
+
+from django.apps import apps
+
+
+@pytest.fixture
+def enable_project_views_sync(settings):
+    settings.PROJECT_VIEWS_SYNC = True
+    apps.get_app_config('projects').ready()
+
+
+@pytest.fixture
+def enable_project_tasks_sync(settings):
+    settings.PROJECT_TASKS_SYNC = True
+    apps.get_app_config('projects').ready()
