+++ rdmo/projects/tests/test_handlers_m2m_views_catalogs.py
@@ -1,55 +1,16 @@
 import pytest
 
-from django.contrib.sites.models import Site
-
 from rdmo.projects.models import Project
-from rdmo.questions.models import Catalog
-from rdmo.views.models import View
 
-from .helpers import assert_all_projects_are_synced_with_instance_m2m_field, enable_project_views_sync  # noqa: F401
-
-P_TITLE = "Sync P{}"  # title of newly created test projects
-one_two_three = (1, 2, 3)  # will be used for creating P1, V1, C1, P2, ..etc.
+from .helpers.arrange_project_views_or_tasks import arrange_projects_catalogs_and_views
+from .helpers.assert_project_views_or_tasks import assert_all_projects_are_synced_with_instance_m2m_field
 
 
 @pytest.mark.django_db
-def test_project_views_sync_when_updating_view_catalogs(settings, enable_project_views_sync):  # noqa:F811
+def test_project_views_sync_when_updating_view_catalogs(settings, enable_project_views_sync):
     assert settings.PROJECT_VIEWS_SYNC
-    site_1 = Site.objects.get(id=1)
 
-    # Arrange: the project, catalog and view objects
-    # all catalogs and views should have available=True
-    # P1, P2, P3
-    P = {
-        n: Project.objects.create(
-            title=P_TITLE.format(n),
-            catalog=Catalog.objects.get(id=n), site=site_1)
-        for n in one_two_three
-    }
-    C = {n: P[n].catalog for n in one_two_three}  # C1, C2, C3
-    V = {n: View.objects.get(id=n) for n in one_two_three}  # V1, V2, V3
-
-    # Arrange the catalogs
-    for catalog in C.values():
-        catalog.available = True
-        catalog.sites.clear()
-        catalog.groups.clear()
-        catalog.save()
-
-    # Set a certain initial state for the project.views
-    # this is a sort of random state
-    P[1].views.set([V[1],V[2],V[3]])
-    P[2].views.clear()
-    P[3].views.set([V[3]])
-
-    # Clear everything to reset state for the views.catalogs
-    # which will also affect the project.views
-    for n, view in V.items():
-        view.available = True
-        view.sites.clear()
-        view.groups.clear()
-        view.catalogs.set([C[n]])
-        view.save()
+    P, C, V = arrange_projects_catalogs_and_views()
 
     # === Initial state ===
     # P1 (with C1) has V1, etc..
@@ -64,7 +25,7 @@ def test_project_views_sync_when_updating_view_catalogs(settings, enable_project
     assert set(P[3].views.all()) == {V[3], V[1]}
     # additionally, all of the projects should have T1
     assert Project.objects.filter(views=V[1]).count() == Project.objects.all().count()
-    assert_all_projects_are_synced_with_instance_m2m_field(V[1])
+    assert_all_projects_are_synced_with_instance_m2m_field(V[1], 'catalogs')
 
     # === Update: (from empty) add C1 to V1 → it should appear in P1 only again ===
     V[1].catalogs.add(C[1])  # V1 → [C1]
@@ -72,7 +33,7 @@ def test_project_views_sync_when_updating_view_catalogs(settings, enable_project
     assert set(P[1].views.all()) == {V[1]}
     assert set(P[2].views.all()) == {V[2]}
     assert set(P[3].views.all()) == {V[3]}
-    assert_all_projects_are_synced_with_instance_m2m_field(V[1])
+    assert_all_projects_are_synced_with_instance_m2m_field(V[1], 'catalogs')
 
     # === Update: set [C1, C2] to V1 → it should appear in P1, P2 ===
     V[1].catalogs.set([C[1], C[2]])  # V1 → [C1, C2]
@@ -86,7 +47,7 @@ def test_project_views_sync_when_updating_view_catalogs(settings, enable_project
     assert set(P[1].views.all()) == set()  # removed
     assert set(P[2].views.all()) == {V[2], V[1]}  # stays
     assert set(P[3].views.all()) == {V[3]}
-    assert_all_projects_are_synced_with_instance_m2m_field(V[2])
+    assert_all_projects_are_synced_with_instance_m2m_field(V[2], 'catalogs')
 
     # === Update: remove C2 and add C3 to V1 → it should appear in P1 and P3 ===
     V[1].catalogs.remove(C[2])  # V1 → []
@@ -95,4 +56,4 @@ def test_project_views_sync_when_updating_view_catalogs(settings, enable_project
     assert set(P[1].views.all()) == set()
     assert set(P[2].views.all()) == {V[2]}
     assert set(P[3].views.all()) == {V[3], V[1]}  # got V1
-    assert_all_projects_are_synced_with_instance_m2m_field(V[3])
+    assert_all_projects_are_synced_with_instance_m2m_field(V[3], 'catalogs')
