+++ rdmo/projects/tests/test_view_project.py
@@ -10,7 +10,6 @@ from rdmo.views.models import View
 
 from ..forms import CatalogChoiceField
 from ..models import Project
-from .helpers import enable_project_tasks_sync, enable_project_views_sync  # noqa: F401
 
 users = (
     ('owner', 'owner'),
@@ -467,7 +466,7 @@ def test_project_update_tasks_get(db, client, settings, username, password, proj
             assert response.status_code == 302
 
 
-def test_project_update_tasks_get_not_allowed(db, client, settings, enable_project_tasks_sync):  # noqa:F811
+def test_project_update_tasks_get_not_allowed(db, client, settings, enable_project_tasks_sync):
     assert settings.PROJECT_TASKS_SYNC
     client.login(username='owner', password='owner')
 
@@ -503,7 +502,7 @@ def test_project_update_tasks_post(db, client, settings, username, password, pro
         assert list(Project.objects.get(pk=project_id).tasks.values('id')) == list(project.tasks.values('id'))
 
 
-def test_project_update_tasks_post_not_allowed(db, client, settings, enable_project_tasks_sync):  # noqa:F811
+def test_project_update_tasks_post_not_allowed(db, client, settings, enable_project_tasks_sync):
     assert settings.PROJECT_TASKS_SYNC
     client.login(username='owner', password='owner')
 
@@ -533,7 +532,7 @@ def test_project_update_views_get(db, client, settings, username, password, proj
             assert response.status_code == 302
 
 
-def test_project_update_views_get_not_allowed(db, client, settings, enable_project_views_sync):  # noqa:F811
+def test_project_update_views_get_not_allowed(db, client, settings, enable_project_views_sync):
     assert settings.PROJECT_VIEWS_SYNC
     client.login(username='owner', password='owner')
 
@@ -570,7 +569,7 @@ def test_project_update_views_post(db, client, settings, username, password, pro
         assert list(Project.objects.get(pk=project_id).views.values('id')) == list(project.views.values('id'))
 
 
-def test_project_update_views_post_not_allowed(db, client, settings, enable_project_views_sync):  # noqa:F811
+def test_project_update_views_post_not_allowed(db, client, settings, enable_project_views_sync):
     assert settings.PROJECT_VIEWS_SYNC
     client.login(username='owner', password='owner')
