+++ rdmo/projects/tests/test_viewset_project.py
@@ -5,7 +5,6 @@ from django.contrib.sites.models import Site
 from django.urls import reverse
 
 from ..models import Membership, Project, Snapshot, Value, Visibility
-from .helpers import enable_project_views_sync  # noqa: F401
 
 users = (
     ('owner', 'owner'),
@@ -491,7 +490,7 @@ def test_update_parent(db, client, username, password, project_id):
         assert Project.objects.get(pk=project_id).parent == project.parent
 
 
-def test_update_project_views_not_allowed(db, client, settings, enable_project_views_sync):  # noqa:F811
+def test_update_project_views_not_allowed(db, client, settings, enable_project_views_sync):
     assert settings.PROJECT_VIEWS_SYNC
 
     client.login(username='owner', password='owner')
